src/VumbaSoft.AdventureWorks.Controllers/Production/Illustrations/IllustrationsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Production
{
    [Area("Production")]
    public class IllustrationsController : ValidatedController<IIllustrationValidator, IIllustrationService>
    {
        public IllustrationsController(IIllustrationValidator validator, IIllustrationService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(IllustrationView illustration)
        {
            if (!Validator.CanCreate(illustration))
                return View(illustration);

            Service.Create(illustration);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<IllustrationView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<IllustrationView>(id));
        }

        [HttpPost]
        public ActionResult Edit(IllustrationView illustration)
        {
            if (!Validator.CanEdit(illustration))
                return View(illustration);

            Service.Edit(illustration);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<IllustrationView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
